Make helpers true extension methods.
