+++ src/NuGet.Clients/NuGet.PackageManagement.VisualStudio/Handlers/ProjectRetargetingHandler.cs
@@ -4,16 +4,15 @@
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.Diagnostics.CodeAnalysis;
 using System.Globalization;
 using System.Linq;
 using System.Threading;
 using EnvDTE;
+using Microsoft;
 using Microsoft.VisualStudio;
 using Microsoft.VisualStudio.ComponentModelHost;
 using Microsoft.VisualStudio.Shell;
 using Microsoft.VisualStudio.Shell.Interop;
-using NuGet.Packaging.Core;
 using NuGet.ProjectManagement;
 using NuGet.VisualStudio;
 
@@ -72,8 +71,9 @@ namespace NuGet.PackageManagement.VisualStudio
             var vsTrackProjectRetargeting = serviceProvider.GetService(typeof(SVsTrackProjectRetargeting)) as IVsTrackProjectRetargeting;
             if (vsTrackProjectRetargeting != null)
             {
-                _vsMonitorSelection = (IVsMonitorSelection)serviceProvider.GetService(typeof(IVsMonitorSelection));
-                Debug.Assert(_vsMonitorSelection != null);
+                _vsMonitorSelection = serviceProvider.GetService(typeof(IVsMonitorSelection)) as IVsMonitorSelection;
+                Assumes.Present(_vsMonitorSelection);
+
                 _errorListProvider = new ErrorListProvider(serviceProvider);
                 _dte = dte;
                 _solutionManager = solutionManager;
@@ -231,12 +231,12 @@ namespace NuGet.PackageManagement.VisualStudio
             {
                 await NuGetUIThreadHelper.JoinableTaskFactory.SwitchToMainThreadAsync();
 
-                var project = EnvDteProjectInfoUtility.GetActiveProject(_vsMonitorSelection);
+                var project = _vsMonitorSelection.GetActiveProject();
 
                 if (project != null)
                 {
                     _platformRetargetingProject = null;
-                    var frameworkName = EnvDteProjectInfoUtility.GetTargetFrameworkString(project);
+                    var frameworkName = project.GetTargetFrameworkString();
                     if (NETCore45.Equals(frameworkName, StringComparison.OrdinalIgnoreCase) || Windows80.Equals(frameworkName, StringComparison.OrdinalIgnoreCase))
                     {
                         _platformRetargetingProject = project.UniqueName;
@@ -268,7 +268,7 @@ namespace NuGet.PackageManagement.VisualStudio
 
                             if (ProjectRetargetingUtility.IsProjectRetargetable(nuGetProject))
                             {
-                                var frameworkName = EnvDteProjectInfoUtility.GetTargetFrameworkString(project);
+                                var frameworkName = project.GetTargetFrameworkString();
                                 if (NETCore451.Equals(frameworkName, StringComparison.OrdinalIgnoreCase) || Windows81.Equals(frameworkName, StringComparison.OrdinalIgnoreCase))
                                 {
                                     var packagesToBeReinstalled = await ProjectRetargetingUtility.GetPackagesToBeReinstalled(nuGetProject);
