+++ src/NuGet.Clients/NuGet.PackageManagement.VisualStudio/ProjectSystems/VsMSBuildProjectSystem.cs
@@ -210,8 +210,8 @@ namespace NuGet.PackageManagement.VisualStudio
                 && !fileExistsInProject
                 && !fileName.Equals(ProjectManagement.Constants.PackageReferenceFile)
                 && !fileName.Equals("packages." + ProjectName + ".config")
-                && !fileName.Equals(EnvDteProjectInfoUtility.WebConfig)
-                && !fileName.Equals(EnvDteProjectInfoUtility.AppConfig)
+                && !fileName.Equals(EnvDteProjectExtensions.WebConfig)
+                && !fileName.Equals(EnvDteProjectExtensions.AppConfig)
                 && !fileName.Equals(Path.GetFileName(lockFileFullPath))
                 )
             {
