+++ src/NuGet.Clients/NuGet.PackageManagement.VisualStudio/Utility/NuGetProjectUpgradeUtility.cs
@@ -87,7 +87,7 @@ namespace NuGet.PackageManagement.VisualStudio
         {
             var solutionManager = ServiceLocator.GetInstance<IVsSolutionManager>();
 
-            var projectSafeName = await EnvDteProjectInfoUtility.GetCustomUniqueNameAsync(envDTEProject);
+            var projectSafeName = await envDTEProject.GetCustomUniqueNameAsync();
             var nuGetProject = await solutionManager.GetNuGetProjectAsync(projectSafeName);
             return nuGetProject;
         }
