+++ src/NuGet.Clients/NuGet.PackageManagement.VisualStudio/Utility/RuntimeHelpers.cs
@@ -48,7 +48,7 @@ namespace NuGet.PackageManagement.VisualStudio
         private static bool SupportsBindingRedirects(EnvDTE.Project Project)
         {
             return (Project.Kind != null && SupportedProjectTypes.IsSupportedForBindingRedirects(Project.Kind))
-                && !EnvDteProjectInfoUtility.IsWindowsStoreApp(Project);
+                && !Project.IsWindowsStoreApp();
         }
 
         private static Task AddBindingRedirectsAsync(
@@ -152,7 +152,7 @@ namespace NuGet.PackageManagement.VisualStudio
             }
 
             // Create a binding redirect manager over the configuration
-            var manager = new BindingRedirectManager(EnvDteProjectInfoUtility.GetConfigurationFile(vsProjectAdapter.Project), msBuildNuGetProjectSystem);
+            var manager = new BindingRedirectManager(vsProjectAdapter.Project.GetConfigurationFile(), msBuildNuGetProjectSystem);
 
             // Add the redirects
             manager.AddBindingRedirects(redirects);
