+++ src/NuGet.Clients/NuGet.VisualStudio.Common/GlobalSuppressions.cs
@@ -16,28 +16,13 @@ using System.Diagnostics.CodeAnalysis;
 [assembly: SuppressMessage("Build", "CA1062:In externally visible method 'bool AssemblyBinding.Equals(AssemblyBinding other)', validate parameter 'other' is non-null before using it. If appropriate, throw an ArgumentNullException when the argument is null or add a Code Contract precondition asserting non-null argument.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.AssemblyBinding.Equals(NuGet.VisualStudio.AssemblyBinding)~System.Boolean")]
 [assembly: SuppressMessage("Build", "CA1507:Use nameof in place of string literal 'dependentAssembly'", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.AssemblyBinding.Parse(System.Xml.Linq.XContainer)~NuGet.VisualStudio.AssemblyBinding")]
 [assembly: SuppressMessage("Build", "CA1305:The behavior of 'string.Format(string, object, object)' could vary based on the current user's locale settings. Replace this call in 'ErrorListTableEntry.TryGetValue(string, out object)' with a call to 'string.Format(IFormatProvider, string, params object[])'.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.Common.ErrorListTableEntry.TryGetValue(System.String,System.Object@)~System.Boolean")]
-[assembly: SuppressMessage("Build", "CA1062:In externally visible method 'string EnvDTEExtensions.GetFullVsVersionString(DTE dte)', validate parameter 'dte' is non-null before using it. If appropriate, throw an ArgumentNullException when the argument is null or add a Code Contract precondition asserting non-null argument.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.EnvDTEExtensions.GetFullVsVersionString(EnvDTE.DTE)~System.String")]
-[assembly: SuppressMessage("Build", "CA1062:In externally visible method 'string EnvDTEExtensions.GetSKU(DTE dte)', validate parameter 'dte' is non-null before using it. If appropriate, throw an ArgumentNullException when the argument is null or add a Code Contract precondition asserting non-null argument.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.EnvDTEExtensions.GetSKU(EnvDTE.DTE)~System.String")]
-[assembly: SuppressMessage("Build", "CA1062:In externally visible method 'Project EnvDTEProjectInfoUtility.GetActiveProject(IVsMonitorSelection vsMonitorSelection)', validate parameter 'vsMonitorSelection' is non-null before using it. If appropriate, throw an ArgumentNullException when the argument is null or add a Code Contract precondition asserting non-null argument.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.EnvDteProjectInfoUtility.GetActiveProject(Microsoft.VisualStudio.Shell.Interop.IVsMonitorSelection)~EnvDTE.Project")]
-[assembly: SuppressMessage("Build", "CA1062:In externally visible method 'Task<string> EnvDTEProjectInfoUtility.GetCustomUniqueNameAsync(Project envDTEProject)', validate parameter 'envDTEProject' is non-null before using it. If appropriate, throw an ArgumentNullException when the argument is null or add a Code Contract precondition asserting non-null argument.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.EnvDteProjectInfoUtility.GetCustomUniqueNameAsync(EnvDTE.Project)~System.Threading.Tasks.Task{System.String}")]
-[assembly: SuppressMessage("Build", "CA1062:In externally visible method 'string EnvDTEProjectInfoUtility.GetFullPath(Project envDTEProject)', validate parameter 'envDTEProject' is non-null before using it. If appropriate, throw an ArgumentNullException when the argument is null or add a Code Contract precondition asserting non-null argument.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.EnvDteProjectInfoUtility.GetFullPath(EnvDTE.Project)~System.String")]
-[assembly: SuppressMessage("Build", "CA1062:In externally visible method 'string EnvDTEProjectInfoUtility.GetFullProjectPath(Project envDTEProject)', validate parameter 'envDTEProject' is non-null before using it. If appropriate, throw an ArgumentNullException when the argument is null or add a Code Contract precondition asserting non-null argument.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.EnvDteProjectInfoUtility.GetFullProjectPath(EnvDTE.Project)~System.String")]
-[assembly: SuppressMessage("Build", "CA1062:In externally visible method 'string EnvDTEProjectInfoUtility.GetName(Project envDTEProject)', validate parameter 'envDTEProject' is non-null before using it. If appropriate, throw an ArgumentNullException when the argument is null or add a Code Contract precondition asserting non-null argument.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.EnvDteProjectInfoUtility.GetName(EnvDTE.Project)~System.String")]
-[assembly: SuppressMessage("Build", "CA1062:In externally visible method 'string EnvDTEProjectInfoUtility.GetUniqueName(Project envDTEProject)', validate parameter 'envDTEProject' is non-null before using it. If appropriate, throw an ArgumentNullException when the argument is null or add a Code Contract precondition asserting non-null argument.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.EnvDteProjectInfoUtility.GetUniqueName(EnvDTE.Project)~System.String")]
-[assembly: SuppressMessage("Build", "CA1062:In externally visible method 'bool EnvDTEProjectInfoUtility.IsUnloaded(Project envDTEProject)', validate parameter 'envDTEProject' is non-null before using it. If appropriate, throw an ArgumentNullException when the argument is null or add a Code Contract precondition asserting non-null argument.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.EnvDteProjectInfoUtility.IsUnloaded(EnvDTE.Project)~System.Boolean")]
-[assembly: SuppressMessage("Build", "CA1062:In externally visible method 'bool EnvDTEProjectInfoUtility.IsWebSite(Project envDTEProject)', validate parameter 'envDTEProject' is non-null before using it. If appropriate, throw an ArgumentNullException when the argument is null or add a Code Contract precondition asserting non-null argument.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.EnvDteProjectInfoUtility.IsWebSite(EnvDTE.Project)~System.Boolean")]
-[assembly: SuppressMessage("Build", "CA1308:In method 'GetHashCode', replace the call to 'ToLowerInvariant' with 'ToUpperInvariant'.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.EnvDteProjectInfoUtility.PathComparer.GetHashCode(System.String)~System.Int32")]
 [assembly: SuppressMessage("Build", "CA1062:In externally visible method 'void MessageHelper.ShowError(ErrorListProvider errorListProvider, TaskErrorCategory errorCategory, TaskPriority priority, string errorText, IVsHierarchy hierarchyItem)', validate parameter 'errorListProvider' is non-null before using it. If appropriate, throw an ArgumentNullException when the argument is null or add a Code Contract precondition asserting non-null argument.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.MessageHelper.ShowError(Microsoft.VisualStudio.Shell.ErrorListProvider,Microsoft.VisualStudio.Shell.TaskErrorCategory,Microsoft.VisualStudio.Shell.TaskPriority,System.String,Microsoft.VisualStudio.Shell.Interop.IVsHierarchy)")]
 [assembly: SuppressMessage("Build", "CA1507:Use nameof in place of string literal 'path'", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.PathHelper.EscapePSPath(System.String)~System.String")]
 [assembly: SuppressMessage("Build", "CA1507:Use nameof in place of string literal 'maxWidth'", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.PathHelper.SmartTruncate(System.String,System.Int32)~System.String")]
 [assembly: SuppressMessage("Build", "CA1062:In externally visible method 'Task ProjectHelper.DoWorkInWriterLockAsync(Project project, IVsHierarchy hierarchy, Action<Project> action)', validate parameter 'action' is non-null before using it. If appropriate, throw an ArgumentNullException when the argument is null or add a Code Contract precondition asserting non-null argument.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.ProjectHelper.DoWorkInWriterLockAsync(EnvDTE.Project,Microsoft.VisualStudio.Shell.Interop.IVsHierarchy,System.Action{Microsoft.Build.Evaluation.Project})~System.Threading.Tasks.Task")]
-[assembly: SuppressMessage("Build", "CA1062:In externally visible method 'Task<ProjectNames> ProjectNames.FromDTEProjectAsync(Project dteProject)', validate parameter 'dteProject' is non-null before using it. If appropriate, throw an ArgumentNullException when the argument is null or add a Code Contract precondition asserting non-null argument.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.ProjectNames.FromDTEProjectAsync(EnvDTE.Project)~System.Threading.Tasks.Task{NuGet.VisualStudio.ProjectNames}")]
-[assembly: SuppressMessage("Build", "CA1062:In externally visible method 'RestoreTelemetryEvent.RestoreTelemetryEvent(string operationId, string[] projectIds, RestoreOperationSource source, DateTimeOffset startTime, NuGetOperationStatus status, int packageCount, int noOpProjectsCount, DateTimeOffset endTime, double duration, IntervalTracker intervalTimingTracker)', validate parameter 'intervalTimingTracker' is non-null before using it. If appropriate, throw an ArgumentNullException when the argument is null or add a Code Contract precondition asserting non-null argument.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.RestoreTelemetryEvent.#ctor(System.String,System.String[],NuGet.VisualStudio.RestoreOperationSource,System.DateTimeOffset,NuGet.Common.NuGetOperationStatus,System.Int32,System.Int32,System.DateTimeOffset,System.Double,NuGet.VisualStudio.IntervalTracker)")]
 [assembly: SuppressMessage("Build", "CA1031:Modify 'GetInstanceSafe' to catch a more specific allowed exception type, or rethrow the exception.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.ServiceLocator.GetInstanceSafe``1~``0")]
 [assembly: SuppressMessage("Build", "CA1062:In externally visible method 'TService ServiceProviderExtensions.GetService<TService>(IServiceProvider serviceProvider)', validate parameter 'serviceProvider' is non-null before using it. If appropriate, throw an ArgumentNullException when the argument is null or add a Code Contract precondition asserting non-null argument.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.ServiceProviderExtensions.GetService``1(System.IServiceProvider)~``0")]
-[assembly: SuppressMessage("Build", "CA1062:In externally visible method 'TInterface ServiceProviderExtensions.GetService<TService, TInterface>(IServiceProvider serviceProvider)', validate parameter 'serviceProvider' is non-null before using it. If appropriate, throw an ArgumentNullException when the argument is null or add a Code Contract precondition asserting non-null argument.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.ServiceProviderExtensions.GetService``2(System.IServiceProvider)~``1")]
 [assembly: SuppressMessage("Build", "CA1062:In externally visible method 'Task<TService> ServiceProviderExtensions.GetServiceAsync<TService>(IAsyncServiceProvider site)', validate parameter 'site' is non-null before using it. If appropriate, throw an ArgumentNullException when the argument is null or add a Code Contract precondition asserting non-null argument.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.ServiceProviderExtensions.GetServiceAsync``1(Microsoft.VisualStudio.Shell.IAsyncServiceProvider)~System.Threading.Tasks.Task{``0}")]
-[assembly: SuppressMessage("Build", "CA1062:In externally visible method 'Task<TInterface> ServiceProviderExtensions.GetServiceAsync<TService, TInterface>(IAsyncServiceProvider site)', validate parameter 'site' is non-null before using it. If appropriate, throw an ArgumentNullException when the argument is null or add a Code Contract precondition asserting non-null argument.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.ServiceProviderExtensions.GetServiceAsync``2(Microsoft.VisualStudio.Shell.IAsyncServiceProvider)~System.Threading.Tasks.Task{``1}")]
 [assembly: SuppressMessage("Build", "CA1062:In externally visible method 'void SolutionEventsListener.Advise(IVsSolution vsSolution)', validate parameter 'vsSolution' is non-null before using it. If appropriate, throw an ArgumentNullException when the argument is null or add a Code Contract precondition asserting non-null argument.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.SolutionEventsListener.Advise(Microsoft.VisualStudio.Shell.Interop.IVsSolution)")]
 [assembly: SuppressMessage("Build", "CA1303:Method 'string PackageSourceTelemetry.GetActionName(TelemetryAction action)' passes a literal string as parameter 'message' of a call to 'ArgumentException.ArgumentException(string message, string paramName)'. Retrieve the following string(s) from a resource table instead: \"Unknown value of TelemetryAction\".", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.Telemetry.PackageSourceTelemetry.GetActionName(NuGet.VisualStudio.Telemetry.PackageSourceTelemetry.TelemetryAction)~System.String")]
 [assembly: SuppressMessage("Build", "CA1305:The behavior of 'int.ToString()' could vary based on the current user's locale settings. Replace this call in 'PackageSourceTelemetry.ToStatusCodeTelemetry(Dictionary<int, int>)' with a call to 'int.ToString(IFormatProvider)'.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.Telemetry.PackageSourceTelemetry.ToStatusCodeTelemetry(System.Collections.Generic.Dictionary{System.Int32,System.Int32})~NuGet.Common.TelemetryEvent")]
@@ -46,7 +31,6 @@ using System.Diagnostics.CodeAnalysis;
 [assembly: SuppressMessage("Build", "CA1825:Avoid unnecessary zero-length array allocations.  Use Array.Empty<VsHierarchyItem>() instead.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.VsHierarchyUtility.GetExpandedProjectHierarchyItems(EnvDTE.Project)~System.Collections.Generic.ICollection{NuGet.VisualStudio.VsHierarchyItem}")]
 [assembly: SuppressMessage("Build", "CA1062:In externally visible method 'Project VsHierarchyUtility.GetProjectFromHierarchy(IVsHierarchy pHierarchy)', validate parameter 'pHierarchy' is non-null before using it. If appropriate, throw an ArgumentNullException when the argument is null or add a Code Contract precondition asserting non-null argument.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.VsHierarchyUtility.GetProjectFromHierarchy(Microsoft.VisualStudio.Shell.Interop.IVsHierarchy)~EnvDTE.Project")]
 [assembly: SuppressMessage("Build", "CA1062:In externally visible method 'string VsHierarchyUtility.GetProjectPath(IVsHierarchy project)', validate parameter 'project' is non-null before using it. If appropriate, throw an ArgumentNullException when the argument is null or add a Code Contract precondition asserting non-null argument.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.VsHierarchyUtility.GetProjectPath(Microsoft.VisualStudio.Shell.Interop.IVsHierarchy)~System.String")]
-[assembly: SuppressMessage("Build", "CA1062:In externally visible method 'string[] VsHierarchyUtility.GetProjectTypeGuids(Project project)', validate parameter 'project' is non-null before using it. If appropriate, throw an ArgumentNullException when the argument is null or add a Code Contract precondition asserting non-null argument.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.VsHierarchyUtility.GetProjectTypeGuids(EnvDTE.Project)~System.String[]")]
 [assembly: SuppressMessage("Build", "CA1062:In externally visible method 'void WindowFrameHelper.AddF1HelpKeyword(IVsWindowFrame windowFrame, string keywordValue)', validate parameter 'windowFrame' is non-null before using it. If appropriate, throw an ArgumentNullException when the argument is null or add a Code Contract precondition asserting non-null argument.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.WindowFrameHelper.AddF1HelpKeyword(Microsoft.VisualStudio.Shell.Interop.IVsWindowFrame,System.String)")]
 [assembly: SuppressMessage("Build", "CA1062:In externally visible method 'void WindowFrameHelper.DisableWindowAutoReopen(IVsWindowFrame windowFrame)', validate parameter 'windowFrame' is non-null before using it. If appropriate, throw an ArgumentNullException when the argument is null or add a Code Contract precondition asserting non-null argument.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.WindowFrameHelper.DisableWindowAutoReopen(Microsoft.VisualStudio.Shell.Interop.IVsWindowFrame)")]
 [assembly: SuppressMessage("Build", "CA1305:The behavior of 'string.Format(string, params object[])' could vary based on the current user's locale settings. Replace this call in 'PackageManagementFormat.PackageFormatSelectorLabel.get' with a call to 'string.Format(IFormatProvider, string, params object[])'.", Justification = "<Pending>", Scope = "member", Target = "~P:NuGet.VisualStudio.PackageManagementFormat.PackageFormatSelectorLabel")]
@@ -55,6 +39,5 @@ using System.Diagnostics.CodeAnalysis;
 [assembly: SuppressMessage("Build", "CA1816:Change OutputConsoleLogger.Dispose() to call GC.SuppressFinalize(object). This will prevent derived types that introduce a finalizer from needing to re-implement 'IDisposable' to call it.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.Common.OutputConsoleLogger.Dispose")]
 [assembly: SuppressMessage("Build", "CA1063:Modify 'OutputConsoleLogger.Dispose' so that it calls Dispose(true), then calls GC.SuppressFinalize on the current object instance ('this' or 'Me' in Visual Basic), and then returns.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.Common.OutputConsoleLogger.Dispose")]
 [assembly: SuppressMessage("Build", "CA1062:In externally visible method 'void OutputConsoleLogger.Log(ILogMessage message)', validate parameter 'message' is non-null before using it. If appropriate, throw an ArgumentNullException when the argument is null or add a Code Contract precondition asserting non-null argument.", Justification = "<Pending>", Scope = "member", Target = "~M:NuGet.VisualStudio.Common.OutputConsoleLogger.Log(NuGet.Common.ILogMessage)")]
-[assembly: SuppressMessage("Build", "CA1823:Unused field 'LogEntrySource'.", Justification = "<Pending>", Scope = "member", Target = "~F:NuGet.VisualStudio.Common.OutputConsoleLogger.LogEntrySource")]
 [assembly: SuppressMessage("Build", "CA1063:Provide an overridable implementation of Dispose(bool) on 'OutputConsoleLogger' or mark the type as sealed. A call to Dispose(false) should only clean up native resources. A call to Dispose(true) should clean up both managed and native resources.", Justification = "<Pending>", Scope = "type", Target = "~T:NuGet.VisualStudio.Common.OutputConsoleLogger")]
 [assembly: SuppressMessage("Globalization", "CA1308:Normalize strings to uppercase", Justification = "Need to unify event names to be same as ones produced from telemetry.", Scope = "member", Target = "~M:NuGet.VisualStudio.Telemetry.EtwLogActivity.#ctor(System.String)")]
