+++ src/NuGet.Clients/NuGet.VisualStudio.Common/IDE/EnvDteExtensions.cs
@@ -5,14 +5,14 @@ using System;
 using Microsoft.VisualStudio.Shell;
 using NuGet.VisualStudio.Common;
 
+#pragma warning disable CA1062 // Validate arguments of public methods
+
 namespace NuGet.VisualStudio
 {
     public static class EnvDteExtensions
     {
         public static string GetSKU(this EnvDTE.DTE dte)
         {
-            Verify.ArgumentIsNotNull(dte, nameof(dte));
-
             ThreadHelper.ThrowIfNotOnUIThread();
 
             var sku = dte.Edition;
@@ -30,8 +30,6 @@ namespace NuGet.VisualStudio
 
         public static string GetFullVsVersionString(this EnvDTE.DTE dte)
         {
-            Verify.ArgumentIsNotNull(dte, nameof(dte));
-
             ThreadHelper.ThrowIfNotOnUIThread();
 
             return dte.Edition + "/" + dte.Version;
